requirements.txt
sanic
sanic-ext
python-dotenv
cryptography
flake8
pyjwt
tortoise-orm
aiomysql